Split widget: glass input
