+++ extension/devtools/app/lib/main.dart
@@ -21,6 +21,7 @@ part 'hooks/use_ui_state.dart';
 part 'utils/helpers.dart';
 part 'widgets/action_pill.dart';
 part 'widgets/filter_chip.dart';
+part 'widgets/glass_input.dart';
 part 'widgets/status_badge.dart';
 
 void main() {
@@ -3177,38 +3178,6 @@ class _SignalDetail extends StatelessWidget {
             const SizedBox(height: 12),
             Text(entry!.note, style: textTheme.bodySmall),
           ],
-        ],
-      ),
-    );
-  }
-}
-
-class _GlassInput extends StatelessWidget {
-  const _GlassInput({required this.controller, required this.hintText});
-
-  final TextEditingController controller;
-  final String hintText;
-
-  @override
-  Widget build(BuildContext context) {
-    return _GlassCard(
-      padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 6),
-      child: Row(
-        children: [
-          const Icon(Icons.search, size: 18),
-          const SizedBox(width: 8),
-          Expanded(
-            child: TextField(
-              controller: controller,
-              decoration: InputDecoration(
-                hintText: hintText,
-                hintStyle: Theme.of(context).textTheme.bodySmall,
-                border: InputBorder.none,
-                isDense: true,
-              ),
-              style: Theme.of(context).textTheme.bodyMedium,
-            ),
-          ),
         ],
       ),
     );
